fix(control-item): fixes function that gets the first id
- Refactors to TS and more func style
